Extracting and isolating functions to simplify readability and testability
